[bugfix][hgraph] fix core on large dataset build (#588)

+++ src/data_cell/extra_info_datacell.h
@@ -67,7 +67,8 @@ public:
             return;
         }
         this->max_capacity_ = new_capacity;
-        uint64_t io_size = new_capacity * extra_info_size_;
+        uint64_t io_size =
+            static_cast<uint64_t>(new_capacity) * static_cast<uint64_t>(extra_info_size_);
         uint8_t end_flag =
             127;  // the value is meaningless, only to occupy the position for io allocate
         this->io_->Write(&end_flag, 1, io_size);
